Release 2.0.0 - finally
After a long journey of debugging, it turns out that the problem
is that 'git shortlog --summary --email' does not display any
authors anymore since today, when run in GitHub Actions.

The fix for now is to make 'make authors' tolerant against that,
by dececting there are no authors and then leaving the AUTHORS.md
file unchanged.

Signed-off-by: Andreas Maier <maiera@de.ibm.com>